feat: restructuring project and clean unused files and components
