+++ src/components/Headers.tsx
@@ -1,14 +1,11 @@
-import Logo from "/icon.png";
+import "../index.css"
 import { useState } from "react";
 import "../index.css"
 import { Link, Route, Routes } from "react-router-dom";
-import GetReservation from "./GetReservation";
-import CreateReservation from "./CreateReservation";
-import UpdateReservation from "./UpdateReservation";
-import DeleteReservation from "./DeleteReservation";
 import { Menu } from "lucide-react";
+import { CreateReservation, DeleteReservation, GetReservation, UpdateReservation } from "./Reservations";
 
-export default function HeaderReservation() {
+function HeaderReservation() {
 
     const [menuMobile, setMenuMobile] = useState(false);
 
@@ -31,7 +28,7 @@ export default function HeaderReservation() {
 
                     <Link to="/home">
                         <button className="flex items-center gap-2 cursor-pointer hover:scale-105 duration-100 ease-in-out">
-                            <img src={Logo} alt="Logo" className="w-12 h-12" />
+                            <img src='/icon.png' alt="Logo" className="w-12 h-12" />
                             <h1 className="text-2xl font-bold">Bite & Enjoy</h1>
                         </button>
                     </Link>
@@ -53,9 +50,9 @@ export default function HeaderReservation() {
 
                     <div className="flex items-center gap-4 justify-between w-full">
 
-                        <Link  to="/home">
+                        <Link to="/home">
                             <button className="flex items-center gap-2 cursor-pointer hover:scale-105 duration-100 ease-in-out">
-                                <img src={Logo} alt="Logo" className="w-12 h-12" />
+                                <img src='/icon.png' alt="Logo" className="w-12 h-12" />
                                 <h1 className="text-2xl font-bold">Bite & Enjoy</h1>
                             </button>
                         </Link>
@@ -101,6 +98,65 @@ export default function HeaderReservation() {
                 <Route path="/delete" element={<DeleteReservation />} />
             </Routes>
         </>
-
     )
 }
+
+function HeaderHomePage() {
+    const [menuMobile, setMenuMobile] = useState(false);
+    const toggleMenu = () => {
+        setMenuMobile(!menuMobile);
+    };
+
+    return (
+        <>
+            <header id="menu" className="w-full justify-center p-4 bg-white shadow-md fixed top-0 z-10">
+                <nav className="header flex flex-wrap justify-center items-center gap-x-40">
+                    <a href="#Hero" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Inicio</a>
+                    <a href="#Menu" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Menú</a>
+                    <a href="#Hero" className="flex items-center gap-2 cursor-pointer hover:scale-105 duration-100 ease-in-out">
+                        <img src='/icon.png' alt="Logo" className="w-12 h-12" />
+                        <h1 className="text-2xl font-bold">Bite & Enjoy</h1>
+                    </a>
+                    <a href="#About" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Nosotros</a>
+
+                    <Link to="/Reservations" target="_blank">
+                        <button className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Reservas</button>
+                    </Link>
+                </nav>
+            </header>
+
+            <div id="menuMobile" className="flex w-full fixed top-0 z-10">
+                <header id="menuMobile" className="w-full flex justify-center p-4 bg-white shadow-md top-0 z-10 fixed">
+                    <div className="flex items-center gap-4 justify-between w-full">
+                        <a href="#" className="flex items-center gap-2 cursor-pointer hover:scale-105 duration-100 ease-in-out">
+                            <img src='/icon.png' alt="Logo" className="w-12 h-12" />
+                            <h1 className="text-2xl font-bold">Bite & Enjoy</h1>
+                        </a>
+                        <button onClick={toggleMenu} id="btnMenuMobile" className="flex text-xl items-center font-semibold cursor-pointer">
+                            <Menu />
+                        </button>
+                    </div>
+                </header>
+
+                {menuMobile && (
+
+                    <div id="menuMobile" className="w-full flex justify-center">
+
+                        <nav id="menuMobile2" className="flex-col gap-6 py-6 gap-x-40 mt-20 border-t-2 duration-300 ease-in-out bg-white shadow-md w-full flex items-center z-10">
+                            <a onClick={toggleMenu} href="#" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Inicio</a>
+                            <a onClick={toggleMenu} href="#Menu" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Menú</a>
+                            <a onClick={toggleMenu} href="#About" className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Nosotros</a>
+                            <Link onClick={toggleMenu} to="/Reservations" target="_blank">
+                                <button className="text-[1.3rem] font-medium hover:text-orange-400 hover:scale-110 duration-200 ease-in-out">Reservas</button>
+                            </Link>
+
+                        </nav>
+                    </div>
+                )}
+            </div>
+
+        </>
+    )
+}
+
+export { HeaderHomePage, HeaderReservation }
