+++ src/pages/Homepage.jsx
@@ -1,12 +1,11 @@
 import { useLocation } from 'react-router-dom';
-import Header from '../components/Header';
-import ReservationHeader from '../components/HeaderReservation';
-import ContainerMenu from '../components/ContainerMenu';
+import ContainerMenu from '../components/Menu';
 import Footer from '../components/Footer';
 import Clients from '../components/Clients';
 import PostHero from '../components/PostHero';
-import Hero from '../components/Hero';
+import { Hero } from '../components/Hero';
 import AboutUs from '../components/AboutUs';
+import { HeaderHomePage, HeaderReservation } from '../components/Headers';
 
 function HomePage() {
 
@@ -15,7 +14,7 @@ function HomePage() {
 
   return (
     <section className='w-full justify-center items-center flex flex-col bodyHomePage'>
-      {!isReservationsPath ? <Header /> : <ReservationHeader />}
+      {!isReservationsPath ? <HeaderHomePage /> : <HeaderReservation />}
       <div className="bgBodyHomePage"></div>
       <Hero />
       <PostHero />
