+++ src/types/CardDishProps.ts
@@ -1,0 +1,6 @@
+export type CardDishProps = {
+    name: string
+    ingredients: string
+    img: string
+    entradas: string
+}
