+++ src/types/CardReservationProps.ts
@@ -1,0 +1,9 @@
+export type CardReservationProps = {
+    id?: string;
+    name: string;
+    email: string;
+    cedula: string;
+    telefono: string;
+    Npersonas: string;
+    fecha: string;
+};
